jtag: drivers: xlnx-axi-xvc: Add support for Xilinx XVC over direct bus interface (AXI)
This change was derived from the PCIe implementation but is intended to
be used with a direct bus connection to the debug bridge instead of going
across PCIe.

This is useful when using on a Zynq/ZynqMP/uBlaze host with direct
access to the debug bridge over AXI.  You can then use the debug bridge
Xilinx IP (AXIXVC) to debug a remote device.

The Xilinx XVC interface has no srst/trst signal support.  That is added
via libgpiod.

Signed-off-by: Jeremy Garff <jeremy.garff@wdc.com>
Change-Id: I11264b591cbf315a1bc7c9fb1ffe470b449da423